# api.yaml
openapi: 3.0.0
info:
  version: '0.1'
  title: mentor-signal
servers:
  - url: 'http://mentor-signal.earthmark.net'
paths:
  /mentee:
    post:
      summary: 'Creates a new ticket for a mentee'
      operationId: createMenteeTicket
      tags:
        - mentee
      requestBody:
        description: 'The content of the ticket to create'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createTicketBody'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/createTicketBody'
      parameters:
        - name: id
          in: query
          description: ID of the user
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: 'Creates a new ticket, returning the ticket id, the default return format is urlencoded.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ticket'
            application/x-www-form-urlencoded:
              schema:
                $ref: '#/components/schemas/ticket'
        '400':
          description: One or more fields was not filled out.
  /mentee/{ticketId}:
    get:
      summary: 'Retrieves the status of a ticket'
      operationId: getMenteeTicket
      tags:
        - mentee
      parameters:
        - name: ticketId
          in: query
          description: ID of the ticket, this is a discord snowflake.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'The ticket, the default return format is urlencoded.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ticket'
            application/x-www-form-urlencoded:
              schema:
                $ref: '#/components/schemas/ticket'
        '404':
          description: The requested ticket did not exist.
  /mentee/{ticketId}/cancel:
    post:
      summary: 'Cancels a ticket if it is in progress'
      operationId: cancelMenteeTicket
      tags:
        - mentee
      parameters:
        - name: ticketId
          in: query
          description: ID of the ticket, this is a discord snowflake.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'The ticket that was canceled, the default return format is urlencoded.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ticket'
            application/x-www-form-urlencoded:
              schema:
                $ref: '#/components/schemas/ticket'
        '404':
          description: The requested ticket did not exist.
        '400':
          description: The requested ticket was already completed or canceled.
components:
  schemas:
    createTicketBody:
      type: object
      required:
        - name
        - lang
        - desc
        - session
      properties:
        name:
          type: string
        lang:
          type: string
        desc:
          type: string
        session:
          type: string
    ticket:
      type: object
      required:
        - ticket
        - status
      properties:
        ticket:
          type: string
        status:
          type: string
        mentor:
          type: string